refactor: move api, storage, platform, session into src/core
