+++ src/core/api/usdb/cover.ts
