+++ src/core/platform.ts
